Moved all pokemon-related types into pokemon.rs
